Refactor SpecFlow service registrations
